work

+++ base/plugins/iscore-plugin-space/Space/square_renderer.hpp
@@ -23,7 +23,7 @@ class square_renderer
                     render_device.add(
                             GiNaC::ex_to<GiNaC::numeric>(map.at(s.variables()[0])).to_double() - side/2.,
                             GiNaC::ex_to<GiNaC::numeric>(map.at(s.variables()[1])).to_double() - side/2.,
-                            side, side);
+                            double(side), double(side));
                 }
             });
